feat(config): change interval to float64 to support sub-minute values (e.g. 0.5 = 30s)

+++ config/config.go
@@ -8,16 +8,16 @@ import (
 )
 
 const (
-	DefaultInterval  = 10
+	DefaultInterval  = 10.0
 	DefaultMaxDelays = 3
 	DefaultModel     = "deepseek/deepseek-v4-flash"
 	FileName         = "quill.toml"
 )
 
 type Config struct {
-	Interval  int    `toml:"interval"`
-	MaxDelays int    `toml:"max_delays"`
-	Model     string `toml:"model"`
+	Interval  float64 `toml:"interval"`
+	MaxDelays int     `toml:"max_delays"`
+	Model     string  `toml:"model"`
 }
 
 func Default() Config {
